Make a component out of the Search form and result list
